Add disable-chunked-transfer to conformance app.
Unfortunately Azure Functions has a bug that doesn't allow it to work with chunked requests right now. https://github.com/Azure/azure-functions-host/issues/4926

This flag made the "usage" output pretty ugly because it is so long, so I opted to leave it out (similar to "--serializer" which I'm not sure was an intentional ommission or not). Totally open to another way!